Reorganise into independant workspaces
This was started so that the python library could be a stand alone
package, so it can be compiled as a separate library type to the
remainder of the package. This allows tests to be run without
any jank. This will also help when/if the project grows in size.